src/__init__.py
""" Initialize the Flask app. """

from dotenv import load_dotenv
import os
from flask import Flask
from flask_cors import CORS
from flask_sqlalchemy import SQLAlchemy
from flask_migrate import Migrate
from flask_jwt_extended import JWTManager
from src.config import config_by_name

# Load environment variables from .env file
load_dotenv()

cors = CORS()
db = SQLAlchemy()
migrate = Migrate()
jwt = JWTManager()

# Ensure the app is configured and SQLAlchemy is initialized before importing models
import src.models.base
import src.models.amenity
import src.models.city
import src.models.country
import src.models.place
import src.models.review
import src.models.user

def create_app(config_name=None) -> Flask:
    """
    Create a Flask app with the given configuration class.
    The default configuration class is DevelopmentConfig.
    """
    app = Flask(__name__)
    app.url_map.strict_slashes = False

    if config_name is None:
        config_name = os.getenv('FLASK_CONFIG', 'development')

    app.config.from_object(config_by_name[config_name])

    print(f"Running with configuration: {config_name}")
    print(f"USE_DATABASE is set to: {app.config['USE_DATABASE']}")

    register_extensions(app)
    register_routes(app)
    register_handlers(app)

    return app


def register_extensions(app: Flask) -> None:
    """Register the extensions for the Flask app."""
    cors.init_app(app, resources={r"/api/*": {"origins": "*"}})
    db.init_app(app)
    migrate.init_app(app, db)
    jwt.init_app(app)


def register_routes(app: Flask) -> None:
    """Import and register the routes for the Flask app."""
    from src.routes.users import users_bp
    from src.routes.countries import countries_bp
    from src.routes.cities import cities_bp
    from src.routes.places import places_bp
    from src.routes.amenities import amenities_bp
    from src.routes.reviews import reviews_bp

    app.register_blueprint(users_bp)
    app.register_blueprint(countries_bp)
    app.register_blueprint(cities_bp)
    app.register_blueprint(places_bp)
    app.register_blueprint(reviews_bp)
    app.register_blueprint(amenities_bp)


def register_handlers(app: Flask) -> None:
    """Register the error handlers for the Flask app."""
    app.errorhandler(404)(lambda e: (
        {"error": "Not found", "message": str(e)}, 404
    ))
    app.errorhandler(400)(
        lambda e: (
            {"error": "Bad request", "message": str(e)}, 400
        )
    )


if __name__ == "__main__":
    app = create_app()
    app.run(debug=True)