requirements.txt
flask
flask-cors
gunicorn
SQLAlchemy
Flask-SQLAlchemy
Flask-Migrate
Flask-JWT-Extended
python-dotenv
psycopg2-binary